requirements.txt
discord.py[voice]~=2.4.0
patool~=3.1.0
yt_dlp~=2024.12.13
